+++ src/components/formButtons/__tests__/formButtons.test.tsx
@@ -1,18 +1,10 @@
 import React from "react";
 import renderer from "react-test-renderer";
 import { FormButtons } from "../FormButtons";
+import { defaultSettings } from "@src/shared/settings";
 
 const mockFunction = jest.fn();
-const settings = {
-    messageMaxWidthStyle: "",
-    messageColorUserStyle: "",
-    messageColorNonUserStyle: "",
-    messagePaddingStyle: "",
-    messageBorderRadiusStyle: "",
-    inputBoxMaxWidthStyle: "",
-    textColorUserStyle: "",
-    textColorNonUserStyle: "",
-};
+const settings = { ...defaultSettings };
 
 describe("FormButtons", () => {
     it("renders correctly", () => {
@@ -25,6 +17,7 @@ describe("FormButtons", () => {
                     setLiveSettings={mockFunction}
                     savedSettings={settings}
                     setSavedSettings={mockFunction}
+                    onRestoreDefaults={mockFunction}
                 />,
             )
             .toJSON();